requirements.txt
-f /usr/share/pip-wheels
amqp==2.5.0
asgiref==3.3.1
backcall==0.1.0
backports.csv==1.0.7
beautifulsoup4==4.7.1
billiard==3.6.0.0
cairocffi==1.0.2
CairoSVG==2.3.0
celery==4.3.0
certifi==2019.6.16
cffi==1.12.2
chardet==3.0.4
cssselect2==0.2.1
cssutils==2.2.0
decorator==4.3.0
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.15
django-admin-rangefilter==0.5.4
django-appconf==1.0.3
django-autocomplete-light==3.5.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-filter==2.2.0
django-geojson==2.11.0
django-imagekit==4.0.2
django-import-export==1.2.0
django-inlinecss==0.3.0
django-jet==1.0.7
django-leaflet==0.24.0
django-logentry-admin==1.0.4
django-markdown-deux==1.0.5
django-mathfilters==0.4.0
django-photologue==3.9
django-sortedm2m==1.5.0
django-starcross-gallery==1.0.6
django-taggit==1.3.0
django-tinymce==2.8.0
django-todo==2.0.3
django-weasyprint==0.5.4
django-widget-tweaks==1.4.8
django-xhtml2pdf==0.0.4
djangorestframework==3.11.1
email-reply-parser==0.5.9
et-xmlfile==1.0.1
ExifRead==2.1.2
future==0.18.2
html5lib==1.0.1
idna==2.8
importlib-metadata==4.0.0
ipython==6.5.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.12.1
kombu==4.6.3
lxml==4.3.3
markdown2==2.3.8
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==3.0.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
pilkit==2.0
Pillow==5.0.0
prompt-toolkit==1.0.15
psutil==5.6.3
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.2.0
PyJWT==1.7.1
pynliner==0.8.0
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2018.5
PyYAML==5.1.2
redis==3.2.1
reportlab==3.5.13
requests==2.22.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
soupsieve==1.9.1
sqlparse==0.3.0
tablib==0.13.0
tinycss2==1.0.1
traitlets==4.3.2
twilio==6.55.0
typing-extensions==3.7.4.3
Unidecode==1.0.23
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
WeasyPrint==45
webencodings==0.5.1
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.1
